+++ src/samples/HandlingAttachments/Resources/agents-sdk.png
