+++ src/samples/HandlingAttachments/appsettings.json
@@ -1,17 +1,16 @@
 {
-  "TokenValidation": {
-    "Audiences": [
-      "{{ClientId}}" // this is the Client ID used for the Azure Bot
-    ],
-    "TenantId":  "{{TenantId}}"
+  "AgentApplication": {
+    "StartTypingTimer": false,
+    "RemoveRecipientMention": false,
+    "NormalizeMentions": false
   },
 
   "Connections": {
-    "BotServiceConnection": {
+    "ServiceConnection": {
       "Settings": {
         "AuthType": "ClientSecret", // this is the AuthType for the connection, valid values can be found in Microsoft.Agents.Authentication.Msal.Model.AuthTypes.  The default is ClientSecret.
         "AuthorityEndpoint": "https://login.microsoftonline.com/{{TenantId}}",
-        "ClientId": "{{ClientId}}", // this is the Client ID used for the connection.
+        "ClientId": "{{ClientId}}", // this is the Client ID used for the Azure Bot
         "ClientSecret": "00000000-0000-0000-0000-000000000000", // this is the Client Secret used for the connection.
         "Scopes": [
           "https://api.botframework.com/.default"
@@ -22,7 +21,7 @@
   "ConnectionsMap": [
     {
       "ServiceUrl": "*",
-      "Connection": "BotServiceConnection"
+      "Connection": "ServiceConnection"
     }
   ],
 
@@ -30,7 +29,7 @@
     "LogLevel": {
       "Default": "Information",
       "Microsoft.AspNetCore": "Warning",
-      "Microsoft.Copilot": "Warning",
+      "Microsoft.Agents": "Warning",
       "Microsoft.Hosting.Lifetime": "Information"
     }
   }
