--- src/samples/test-bots/HandlingAttachments/Properties/launchSettings.TEMPLATE.json
@@ -1,16 +1,0 @@
-{
-  "profiles": {
-    "EchoBot": {
-      "commandName": "Project",
-      "launchBrowser": false,
-      "environmentVariables": {
-        "ASPNETCORE_ENVIRONMENT": "Development",
-        "Connections__BotServiceConnection__Settings__TenantId": "<<TENANT_ID>>",
-        "Connections__BotServiceConnection__Settings__ClientId": "<<APP_ID>>",
-        "Connections__BotServiceConnection__Settings__AuthorityEndpoint": "https://login.microsoftonline.com/<<TENANT_ID>>",
-        "Connections__BotServiceConnection__Settings__ClientSecret": "<<CLIENT_SECRET>>"
-      },
-      "applicationUrl": "https://localhost:65284;http://localhost:3978"
-    }
-  }
-}
